Moved homepage-slice-component under pages/home
